FIX: rename `stl` to `stl_file` to  fix circular import error in jupyter nb
